It looks like Synergy uses ISO-8859-1 for stdout.
It looks like Synergy uses ISO-8859-1 for stdout-output, but it is read with the current system-encoding.

If your server uses UTF-8 and the a task-synopsis contains special characters (like german umlauts) you get illegal characters in the synopsis and changelog.
I fixed it by explicitly setting the encoding in Commands.java.

Closes #4