Fixed problem with GetMemberStatusCommand.
Sometimes or on some systems it returns a newline in front of the result. I changed the command to provide a little more robustness about slightly changing output-format.
Additionally I removed the excessive finduse-queries, because the result isn't used anywhere and there can be lot's of finduse-calls (in our project >100).

Closes #5